fixes in chooser controls

+++ app/model/YearCalculator.php
@@ -9,9 +9,9 @@ class YearCalculator extends Object {
     public function getCurrentYear(ModelContest $contest) {
         switch ($contest->contest_id) {
             case ModelContest::ID_FYKOS:
-                return ceil((time() - strtotime('1987-09-01')) / self::YEAR);
+                return (int)ceil((time() - strtotime('1987-09-01')) / self::YEAR);
             case ModelContest::ID_VYFUK:
-                return ceil((time() - strtotime('2011-09-01')) / self::YEAR);
+                return (int)ceil((time() - strtotime('2011-09-01')) / self::YEAR);
             default:
                 return null;
         }
